requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.2
django-environ==0.4.4
djangorestframework==3.7.7
idna==2.6
pkg-resources==0.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
xmltodict==0.11.0
